requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
faiss-cpu>=1.7.4
asyncpg>=0.29.0
psycopg2-binary>=2.9.0